Prevent time variable overflow leading to parameter timeout error (#293)
* Fixed the end_time variable type in node_handle.h. This prevents the erroneous timeout error that is thrown when a board connects to a host running ROS more than 65,535 miliseconds (variable overflow) after the initNode() instruction, in order to get the ROS parameters.

+++ rosserial_client/src/ros_lib/ros/node_handle.h
@@ -494,7 +494,7 @@ namespace ros {
         rosserial_msgs::RequestParamRequest req;
         req.name  = (char*)name;
         publish(TopicInfo::ID_PARAMETER_REQUEST, &req);
-        uint16_t end_time = hardware_.time() + time_out;
+        uint32_t end_time = hardware_.time() + time_out;
         while(!param_recieved ){
           spinOnce();
           if (hardware_.time() > end_time) {
